+++ test/sampler_test.py
@@ -5,6 +5,8 @@ import unittest
 from mock import MagicMock
 import numpy as np
 import os
+import sys
+import shutil
 import copy
 
 
@@ -247,6 +249,53 @@ class TestNestle(unittest.TestCase):
             self.assertDictEqual(expected, self.sampler.kwargs)
 
 
+class TestPolyChord(unittest.TestCase):
+
+    def setUp(self):
+        self.likelihood = MagicMock()
+        self.priors = dict(a=bilby.prior.Uniform(0, 1))
+        self.sampler = bilby.core.sampler.PyPolyChord(self.likelihood, self.priors,
+                                                      outdir='outdir', label='polychord',
+                                                      use_ratio=False, plot=False,
+                                                      skip_import_verification=True)
+
+    def tearDown(self):
+        del self.likelihood
+        del self.priors
+        del self.sampler
+
+    def test_default_kwargs(self):
+        expected = dict(use_polychord_defaults=False, nlive=self.sampler.ndim*25, num_repeats=self.sampler.ndim*5,
+                        nprior=-1, do_clustering=True, feedback=1, precision_criterion=0.001,
+                        logzero=-1e30, max_ndead=-1, boost_posterior=0.0, posteriors=True,
+                        equals=True, cluster_posteriors=True, write_resume=True,
+                        write_paramnames=False, read_resume=True, write_stats=True,
+                        write_live=True, write_dead=True, write_prior=True,
+                        compression_factor=np.exp(-1), base_dir='outdir',
+                        file_root='polychord', seed=-1, grade_dims=list([self.sampler.ndim]),
+                        grade_frac=list([1.0]*len([self.sampler.ndim])), nlives={})
+        self.sampler._setup_dynamic_defaults()
+        self.assertDictEqual(expected, self.sampler.kwargs)
+
+    def test_translate_kwargs(self):
+        expected = dict(use_polychord_defaults=False, nlive=123, num_repeats=self.sampler.ndim*5,
+                        nprior=-1, do_clustering=True, feedback=1, precision_criterion=0.001,
+                        logzero=-1e30, max_ndead=-1, boost_posterior=0.0, posteriors=True,
+                        equals=True, cluster_posteriors=True, write_resume=True,
+                        write_paramnames=False, read_resume=True, write_stats=True,
+                        write_live=True, write_dead=True, write_prior=True,
+                        compression_factor=np.exp(-1), base_dir='outdir',
+                        file_root='polychord', seed=-1, grade_dims=list([self.sampler.ndim]),
+                        grade_frac=list([1.0]*len([self.sampler.ndim])), nlives={})
+        self.sampler._setup_dynamic_defaults()
+        for equiv in bilby.core.sampler.base_sampler.NestedSampler.npoints_equiv_kwargs:
+            new_kwargs = self.sampler.kwargs.copy()
+            del new_kwargs['nlive']
+            new_kwargs[equiv] = 123
+            self.sampler.kwargs = new_kwargs
+            self.assertDictEqual(expected, self.sampler.kwargs)
+
+
 class TestPTEmcee(unittest.TestCase):
 
     def setUp(self):
@@ -397,11 +446,13 @@ class TestRunningSamplers(unittest.TestCase):
 
         self.priors = dict(
             m=bilby.core.prior.Uniform(0, 5), c=bilby.core.prior.Uniform(-2, 2))
+        bilby.core.utils.check_directory_exists_and_if_not_mkdir('outdir')
 
     def tearDown(self):
         del self.likelihood
         del self.priors
         bilby.core.utils.command_line_args.test = False
+        shutil.rmtree('outdir')
 
     def test_run_cpnest(self):
         _ = bilby.run_sampler(
@@ -422,6 +473,11 @@ class TestRunningSamplers(unittest.TestCase):
         _ = bilby.run_sampler(
             likelihood=self.likelihood, priors=self.priors, sampler='nestle',
             nlive=100, save=False)
+
+    def test_run_pypolychord(self):
+        _ = bilby.run_sampler(
+            likelihood=self.likelihood, priors=self.priors,
+            sampler='pypolychord', nlive=100, save=False)
 
     def test_run_ptemcee(self):
         _ = bilby.run_sampler(